requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asgiref==3.10.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.2
build==1.3.0
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
crewai==1.4.1
crewai-tools==1.4.1
cryptography==46.0.3
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker==7.1.0
docstring_parser==0.17.0
drf-yasg==1.21.11
durationpy==0.10
et_xmlfile==2.0.0
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.2
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
json5==0.12.1
json_repair==0.25.2
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
kubernetes==34.1.0
lance-namespace==0.0.20
lance-namespace-urllib3-client==0.0.20
lancedb==0.25.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.21.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
nodeenv==1.9.1
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.7.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
platformdirs==4.5.0
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.39.0
pypdf==6.2.0
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
sse-starlette==3.0.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
tqdm==4.67.1
trello==0.9.7.3
ty==0.0.1a25
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uv==0.9.8
uvicorn==0.38.0
vine==5.1.0
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
youtube-transcript-api==1.2.3
zipp==3.23.0
